req./fix: add requests and fix issue in ExpressionList
